requirements.txt
tensorflow==1.1.0
tflearn==0.3
pillow==4.1.1
colorthief==0.2.1
tqdm==4.11.2
opencv-python==3.2.0.7
h5py==2.7.0
scipy==0.19.0